Remove cmd/kubelet dependency from pkg/kubelet
