+++ cmd/kubelet/app/options/options.go
@@ -34,6 +34,7 @@ import (
 	kubeletscheme "k8s.io/kubernetes/pkg/kubelet/apis/kubeletconfig/scheme"
 	"k8s.io/kubernetes/pkg/kubelet/apis/kubeletconfig/v1alpha1"
 	kubeletconfigvalidation "k8s.io/kubernetes/pkg/kubelet/apis/kubeletconfig/validation"
+	"k8s.io/kubernetes/pkg/kubelet/config"
 	utiltaints "k8s.io/kubernetes/pkg/util/taints"
 )
 
@@ -72,7 +73,7 @@ type KubeletFlags struct {
 	ProviderID string
 
 	// Container-runtime-specific options.
-	ContainerRuntimeOptions
+	config.ContainerRuntimeOptions
 
 	// certDirectory is the directory where the TLS certs are located (by
 	// default /var/run/kubernetes). If tlsCertFile and tlsPrivateKeyFile
