+++ pkg/kubelet/BUILD
@@ -29,7 +29,6 @@ go_library(
         "volume_host.go",
     ],
     deps = [
-        "//cmd/kubelet/app/options:go_default_library",
         "//pkg/api:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/api/v1/helper:go_default_library",
